+++ tests/Integration/Events/BackupHasFailedTest.php
@@ -16,7 +16,7 @@ class BackupHasFailedTest extends TestCase
     /** @test */
     public function it_will_fire_an_event_when_a_backup_has_failed()
     {
-        $this->app['config']->set('laravel-backup.backup.destination.filesystems', [
+        $this->app['config']->set('laravel-backup.backup.destination.disks', [
             'ftp',
         ]);